pyproject.toml
[tool.black]
line-length = 119

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
    "flake8>=2.5.5",
    "isort>=5.11.5",
    "pre-commit>=2.21.0",
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "pytest-django>=4.5.2",
    "bandit>=1.7.5",
    "autopep8>=2.0.4",
]

[tool.pdm.build]
includes = []
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ovino"
version = "0.1.0"
description = "Projeto de ovinocultura (NEPPA - Fabrica de Software)"
authors = [
    {name = "Eduardo da Silva", email = "eduardo.silva@ifc.edu.br"},
]
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "djangorestframework-jwt>=1.11.0",
    "rule>=0.1.1",
    "django-environ>=0.11.2",
    "psycopg2-binary>=2.9.9",
    "django-cors-headers>=4.4.0",
    "django-extensions>=3.2.3",
    "djangorestframework-simplejwt>=5.3.1",
    "rules>=3.4",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"